// API_Client/client/models/create-product-slider-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Online Shop API
 * REST API for the online shop backend
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { ProductSliderTranslationsDto } from './product-slider-translations-dto';

export interface CreateProductSliderDto {
    /**
     * სტაბილური იდენტიფიკატორი (slug), რომლითაც frontend-ი ბლოკს ნებისმიერ გვერდზე embed-ავს — GET /product-sliders/key/:key
     */
    'key': string;
    /**
     * მრავალენოვანი ტექსტი — { ka: {title, viewAllText?}, en?, ru? }, ka სავალდებულოა
     */
    'translations': ProductSliderTranslationsDto;
    /**
     * \"ყველას ნახვა\" ღილაკის ლინკი (მაგ. კონკრეტულ კატეგორიაზე ან ფილტრზე)
     */
    'viewAllLink'?: string;
    /**
     * აქტიურია თუ არა ბლოკი
     */
    'isActive'?: boolean;
    /**
     * დალაგების რიგი
     */
    'sortOrder'?: number;
    /**
     * ბლოკში ჩასასმელი პროდუქტების ID-ები, სასურველი რიგით (მასივის თანმიმდევრობა = ჩვენების რიგი). სურვილისამებრ — შემდეგ ცალკეც შეიძლება PUT /product-sliders/:id/items-ით
     */
    'productIds'?: Set<number>;
}